requirements.txt
fastapi>=0.111.0
email-validator>=2.1.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.29
aiosqlite>=0.20.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.27.0
pytest>=8.2.0
pytest-asyncio>=0.23.6
bcrypt>=4.1.0
boto3>=1.34.0
snowflake-connector-python>=3.10.0
